KJG-70: refactor
